requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Werkzeug
PyMySQL
python-dotenv
mysqlclient